backend/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-environ==0.11.2
django-htmx==1.17.2
django-tailwind==3.8.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==21.2.0
